+++ examples/testbench/src/components/Web3Provider.tsx
@@ -41,6 +41,7 @@ export const ckConfig = getDefaultConfig({
   appName: 'ConnectKit testbench',
   appIcon: '/app.png',
   walletConnectProjectId: process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID!,
+  coinbaseWalletPreference: 'smartWalletOnly',
 });
 const customConfig = {
   ...ckConfig,