+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+
+# x.y.z
+
+This update adds support for [Coinbase Smart Wallet](https://smartwallet.dev) and adds additional support for the latest versions of peer dependencies `wagmi` and `viem`.
+
+## New
+
+- Adds support for [Coinbase Smart Wallet](https://smartwallet.dev).
+
+## Updated
+
+- Changed default setting for `enforceSupportedChains` to `false` to allow for a better default user and developer experience.
+
 # 1.7.3
 
 This update fixes a few bugs and improves the stability of the QR code generation for WalletConnect and Coinbase Wallet.